# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Roil Finance API
  description: |
    Backend API for Roil Finance — an institutional-grade portfolio rebalancing
    platform built on Canton Network with Daml smart contracts.

    Supports portfolio management, DCA scheduling, reward tracking, market data,
    token transfers, yield compounding, and platform administration.
  version: 1.0.0
  contact:
    name: Roil Finance
    url: https://roil.app
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.roil.app
    description: Production

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing `actAs` array of Canton party identifiers.
        Algorithms supported: HS256 (dev), RS256, ES256.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Something went wrong"
...
    AdminStatus:
      type: object
      properties:
        paused:
          type: boolean
        frozen:
          type: boolean
        feeRate:
          type: number
        allowedAssets:
          type: array
          items:
            type: string
        activePortfolios:
          type: integer
        activeDCAs:
          type: integer
        startedAt:
          type: string
          format: date-time
    AuditLogEntry:
      type: object
      properties:
        action:
          type: string
        actor:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: object
...
paths:
...
  /api/admin/status:
    get:
      tags: [Admin]
      summary: Get platform status
      description: Returns current admin state including pause/freeze status, fee rate, and allowed assets.
      responses:
        '200':
          description: Platform status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AdminStatus'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/pause:
    post:
      tags: [Admin]
      summary: Pause all rebalancing and DCA globally
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "Scheduled maintenance"
      responses:
        '200':
          description: Platform paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/resume:
    post:
      tags: [Admin]
      summary: Resume all operations
      responses:
        '200':
          description: Platform resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/fee-rate:
    put:
      tags: [Admin]
      summary: Update platform fee rate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feeRate]
              properties:
                feeRate:
                  type: number
                  minimum: 0
                  maximum: 0.1
                  example: 0.001
      responses:
        '200':
          description: Fee rate updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid fee rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/allowed-assets:
    put:
      tags: [Admin]
      summary: Update allowed asset list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assets]
              properties:
                assets:
                  type: array
                  items:
                    type: string
                  example: ["CC", "USDCx", "CBTC", "ETHx"]
      responses:
        '200':
          description: Allowed assets updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/emergency-freeze:
    post:
      tags: [Admin]
      summary: Emergency freeze all portfolios
      description: |
        Immediately freezes all portfolio operations. This is a safety measure
        that pauses all rebalancing, DCA execution, and compounding.
        Requires manual resume via /api/admin/resume.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "Security incident detected"
      responses:
        '200':
          description: Platform frozen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/audit-log:
    get:
      tags: [Admin]
      summary: Get recent admin actions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'
        '403':
          description: Not an admin party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # System
  # =========================================================================
...